Created Barrier w/ a few bugs.
Currently the zombies get to the barrier and continue moving until they reach the end of the barrier (basically standing on top of it).

This may be because I had to use Float redtangles (RectF) for the barrier while the other class's detection Rectangles are integer-based (Rect).
I don't know for sure that this is the issue, yet.

Minor change to the update method that has to do with the zombies:
    - when the zombie has collided with an active block in the barrier it does not update position. Once the health on the block is
    depleted and the block has therefore been destroyed, the zombie is free to move until it encounters another active block.

-- Note to self, may need to adjust the health on the barrier once it is working as intended. --